Split ObservationModel into abstract class and two subclasses
This makes handling NumericalObservations and CategoricalObservations more explicit.
