+++ src/main/java/model/NumericalObservationModel.java
@@ -1,0 +1,30 @@
+package model;
+
+public class NumericalObservationModel extends AbstractObservationModel {
+
+    private double numericalValue;
+    private String unit;
+
+    public NumericalObservationModel(String observationSystem, String observationCode, double numericalValue, String unit) {
+        this.observationSystem = observationSystem;
+        this.observationCode = observationCode;
+        this.numericalValue = numericalValue;
+        this.unit = unit;
+    }
+
+    public double getNumericalValue() {
+        return numericalValue;
+    }
+
+    public void setNumericalValue(double numericalValue) {
+        this.numericalValue = numericalValue;
+    }
+
+    public String getUnit() {
+        return unit;
+    }
+
+    public void setUnit(String unit) {
+        this.unit = unit;
+    }
+}
