+++ src/main/java/server/Server.java
@@ -6,7 +6,10 @@ import error.DuplicateServerException;
 import kong.unirest.HttpResponse;
 import kong.unirest.JsonNode;
 import kong.unirest.Unirest;
-import model.ObservationModel;
+import model.AbstractObservationModel;
+import model.CategorialObservationModel;
+import model.NumericalObservationModel;
+
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.Enumerations;
 import org.hl7.fhir.r4.model.Observation;
@@ -63,29 +66,22 @@ public class Server {
         return patientID;
     }
 
-    public String createNumericalObservation(ObservationModel observation, String patientID) {
-        Observation fhirObservation = new Observation();
-
-        fhirObservation.getCode().addCoding().setSystem(observation.getObservationSystem()).setCode(observation.getObservationCode());
-        fhirObservation.getSubject().setReference("Patient/" + patientID);
-        fhirObservation.getValueQuantity().setValue(observation.getNumericalValue()).setUnit(observation.getUnit());
-        client.create()
-              .resource(fhirObservation)
-              .execute();
-
-        String observationID = getLastCreatedObservationForPatient(patientID);
-        return observationID;
-    }
-
-    public String createCategoricalObservation(ObservationModel observation, String patientID) {
+    public String createObservation(AbstractObservationModel observation, String patientID) {
         Observation fhirObservation = new Observation();
         fhirObservation.getCode().addCoding().setSystem(observation.getObservationSystem()).setCode(observation.getObservationCode());
         fhirObservation.getSubject().setReference("Patient/" + patientID);
-        fhirObservation.getValueCodeableConcept().addCoding().setSystem(observation.getValueSystem()).setCode(observation.getValueCode());
+        
+        if (observation instanceof NumericalObservationModel) {
+        	NumericalObservationModel numerical = (NumericalObservationModel) observation;
+        	fhirObservation.getValueQuantity().setValue(numerical.getNumericalValue()).setUnit(numerical.getUnit());
+        } else {
+        	CategorialObservationModel categorical = (CategorialObservationModel) observation;
+        	fhirObservation.getValueCodeableConcept().addCoding().setSystem(categorical.getValueSystem()).setCode(categorical.getValueCode());
+        }
+        
         client.create()
               .resource(fhirObservation)
               .execute();
-
         String observationID = getLastCreatedObservationForPatient(patientID);
         return observationID;
     }
@@ -106,7 +102,7 @@ public class Server {
         return allPatientIDs;
     }
 
-    public ArrayList<ObservationModel> getObservationsOfPatient(String patientID) {
+    public ArrayList<AbstractObservationModel> getObservationsOfPatient(String patientID) {
         org.hl7.fhir.r4.model.Bundle results = client
                 .search()
                 .forResource(Observation.class)
@@ -114,7 +110,7 @@ public class Server {
                 .returnBundle(org.hl7.fhir.r4.model.Bundle.class)
                 .execute();
 
-        ArrayList<ObservationModel> allObservations = new ArrayList<>();
+        ArrayList<AbstractObservationModel> allObservations = new ArrayList<>();
         List<Bundle.BundleEntryComponent> entries = results.getEntry();
 
 
@@ -126,14 +122,14 @@ public class Server {
             if (observationResource.hasValueQuantity()) {
                 Double observationValue = observationResource.getValueQuantity().getValue().doubleValue();
                 String observationUnit = observationResource.getValueQuantity().getUnit();
-                ObservationModel observation = new ObservationModel(observationSystem, observationCode, observationValue, observationUnit);
+                NumericalObservationModel observation = new NumericalObservationModel(observationSystem, observationCode, observationValue, observationUnit);
                 observation.setObservationID(observationID);
                 observation.setPatientID(patientID);
                 allObservations.add(observation);
             } else if (observationResource.hasValueCodeableConcept()) {
                 String valueSystem = observationResource.getValueCodeableConcept().getCoding().get(0).getSystem();
                 String valueCode = observationResource.getValueCodeableConcept().getCoding().get(0).getCode();
-                ObservationModel observation = new ObservationModel(observationSystem, observationCode, valueSystem, valueCode);
+                CategorialObservationModel observation = new CategorialObservationModel(observationSystem, observationCode, valueSystem, valueCode);
                 observation.setObservationID(observationID);
                 observation.setPatientID(patientID);
                 allObservations.add(observation);
@@ -143,7 +139,7 @@ public class Server {
     }
 
     public String getLastCreatedObservationForPatient(String patientID) {
-        ArrayList<ObservationModel> allObservations = getObservationsOfPatient(patientID);
+        ArrayList<AbstractObservationModel> allObservations = getObservationsOfPatient(patientID);
         String observationID = allObservations.get(allObservations.size() - 1).getObservationID().split("http://surgipedia.sfb125.de/wiki/Observation/")[1];
         return observationID;
     }
