+++ src/test/java/server/ServerTest.java
@@ -1,6 +1,9 @@
 package server;
 
-import model.ObservationModel;
+import model.AbstractObservationModel;
+import model.CategorialObservationModel;
+import model.NumericalObservationModel;
+
 import org.hl7.fhir.r4.model.Enumerations;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -26,17 +29,17 @@ public class ServerTest {
 
     @Test
     public void testCreateNumericalObs() {
-        ObservationModel observation = new ObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "bilirubin_concentration", 1, "mg/dl");
+        NumericalObservationModel observation = new NumericalObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "bilirubin_concentration", 1, "mg/dl");
         String patientID = server.createPatient("Doe", "John", Enumerations.AdministrativeGender.MALE);
-        String observationID = server.createNumericalObservation(observation, patientID);
+        String observationID = server.createObservation(observation, patientID);
         assertEquals(167, observationID.length());
     }
 
     @Test
     public void testCreateCategoricalObs() {
-        ObservationModel observation = new ObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "lymph_node_staging", "http://sfb125.de/ontology/ihCCApplicationOntology/", "cN1");
+        CategorialObservationModel observation = new CategorialObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "lymph_node_staging", "http://sfb125.de/ontology/ihCCApplicationOntology/", "cN1");
         String patientID = server.createPatient("Doe", "Jane", Enumerations.AdministrativeGender.FEMALE);
-        String observationID = server.createCategoricalObservation(observation, patientID);
+        String observationID = server.createObservation(observation, patientID);
         assertEquals(162, observationID.length());
     }
 
@@ -52,16 +55,16 @@ public class ServerTest {
     @Test
     public void testGetAllObsForPatient() {
         String patientID = server.createPatient("Doe", "Jane", Enumerations.AdministrativeGender.FEMALE);
-        ArrayList<ObservationModel> observations1 = server.getObservationsOfPatient(patientID);
+        ArrayList<AbstractObservationModel> observations1 = server.getObservationsOfPatient(patientID);
         observations1.size();
         assertEquals(1, observations1.size());
-        ObservationModel observationNumerical = new ObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "bilirubin_concentration", 1, "mg/dl");
-        server.createNumericalObservation(observationNumerical, patientID);
-        ArrayList<ObservationModel> observations2 = server.getObservationsOfPatient(patientID);
+        NumericalObservationModel observationNumerical = new NumericalObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "bilirubin_concentration", 1, "mg/dl");
+        server.createObservation(observationNumerical, patientID);
+        ArrayList<AbstractObservationModel> observations2 = server.getObservationsOfPatient(patientID);
         assertEquals(2, observations2.size());
-        ObservationModel observationCategorical = new ObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "lymph_node_staging", "http://sfb125.de/ontology/ihCCApplicationOntology/", "cN1");
-        server.createCategoricalObservation(observationCategorical, patientID);
-        ArrayList<ObservationModel> observations3 = server.getObservationsOfPatient(patientID);
+        CategorialObservationModel observationCategorical = new CategorialObservationModel("http://sfb125.de/ontology/ihCCApplicationOntology/", "lymph_node_staging", "http://sfb125.de/ontology/ihCCApplicationOntology/", "cN1");
+        server.createObservation(observationCategorical, patientID);
+        ArrayList<AbstractObservationModel> observations3 = server.getObservationsOfPatient(patientID);
         assertEquals(3, observations3.size());
     }
 }
